Fixes:#3550 ; Final Review Changes Fixed
