+++ docs/modules/ROOT/pages/kotlinlib/android-examples.adoc
@@ -16,7 +16,7 @@ a starting point for further experimentation and development.
 These are the main Mill Modules that are relevant for building Android apps:
 
 * {mill-doc-url}/api/latest/mill/scalalib/AndroidSdkModule.html[`mill.scalalib.AndroidSdkModule`]: Handles Android SDK management and tools.
-* {mill-doc-url}/api/latest/mill/kotlinlib/android/AndroidAppModule.html[`mill.kotlinlib.android.AndroidAppModule`]: Provides a framework for building Android applications.
+* {mill-doc-url}/api/latest/mill/kotlinlib/android/AndroidAppKotlinModule.html[`mill.kotlinlib.android.AndroidAppKotlinModule`]: Provides a framework for building Android applications.
 * {mill-doc-url}/api/latest/mill/kotlinlib/KotlinModule.html[`mill.kotlinlib.KotlinModule`]: General Kotlin build tasks like compiling Kotlin code and creating JAR files.
 
 == Simple Android Hello World Application
@@ -27,10 +27,10 @@ This example demonstrates how to create a basic "Hello World" Android applicatio
 using the Mill build tool. It outlines the minimum setup required to compile Kotlin code,
 package it into an APK, and run the app on an Android device.
 
-== Understanding `AndroidSdkModule` and `AndroidAppModule`
+== Understanding `AndroidSdkModule` and `AndroidAppKotlinModule`
 
 The two main modules you need to understand when building Android apps with Mill
-are `AndroidSdkModule` and `AndroidAppModule`.
+are `AndroidSdkModule` and `AndroidAppKotlinModule`.
 
 `AndroidSdkModule`:
 
@@ -38,7 +38,7 @@ are `AndroidSdkModule` and `AndroidAppModule`.
 tools like `aapt`, `d8`, `zipalign`, and `apksigner`. These tools are used
 for compiling, packaging, and signing Android applications.
 
-`AndroidAppModule`:
+`AndroidAppKotlinModule`:
 This module provides the step-by-step workflow for building an Android app. It handles
 everything from compiling the code to generating a signed APK for distribution.
